Remove the type option from the usage
The type option is really only needed when running inside an instance.
Hence it shouldn't come up as an option when someone types -h to see
what arguments this script takes.

Also, removed the refernece to reading in the personal.env file when
run from an instance point of view (ie, --type option)