Created a new command line for process_dc_env.py
the script process_dc_env.py was choking on all the extra arguments
that are needed for the calling script but not for process_dc_env.py.
So I constructed the command line from the arguments passed in and only
passed the ones it needs.  Also, remove the comments around the code
that would remove the temporary config that is created for the session.