Catching more edge conditions
If the name is found when trying to set up the IAM user, then we
shouldn't try to create it again.  However, if we don't then we can't
get the keys and if we can't get the keys then we can't set up their AWS
configs.  so it would have to be done manually. UGH.  It would be easier
if they just create a new user.