Refactor the cleanup of the configs
Redirected the output of the upload ssh keys along with the add and
remove user from the group public-key-transfer.  Also, cleaned up the
cleanup AWS configs so that if they had a set of config/credentials
before that it is moved out of place while we do this script.  Then
append what we just did to the originals and move them back.  This way
the original stuff won't be touched but will still be there.  Now it
could happen that the stuff we append might be a duplicate, but it has
an updated set of keys at least.